SUP_2406_github_status_translation
